Special Root Container options factory
